% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_pcard.R
\name{ml_pcard}
\alias{ml_pcard}
\title{ml_pcard An R wrapper for the PCARD Spark Package that performs Random Discretization and PCA, then joins the results and trains an ensemble of decision trees.}
\usage{
ml_pcard(x, response, features = dplyr::tbl_vars(x), numTrees = 10,
  numBins = 5)
}
\arguments{
\item{x}{- a tbl_spark object}

\item{response}{- name of the column containing group labels}

\item{features}{- features to use to predict group membership}

\item{numTrees}{- number of decision trees to use ensemble}

\item{numBins}{- number of bins in random discretization}
}
\value{
A Spark model object that can be used in predict or sdf_predict
}
\description{
ml_pcard An R wrapper for the PCARD Spark Package that performs Random Discretization and PCA, then joins the results and trains an ensemble of decision trees.
}

